Refactor middlewares manga endpoints
