+++ middlewares/manga/controllers/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+  getAll: require('./getAll'),
+  getById: require('./getById'),
+  addManga: require('./addManga'),
+  deleteAllManga: require('./deleteAllManga'),
+  deleteMangaById: require('./deleteMangaById'),
+  updateMangaData: require('./updateMangaData'),
+}
